Cleaned Up File Structure
tidied up the files by combining the playermanager and the compounding manager into one script
