+++ Assets/Scenes/Debug Scene/Scripts/Compounding/MainCompoundingManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8d1d7f42f7566254c8e2c1a870588b84
+guid: 1b7d92570952f944aa157b5284023930
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
