--- Assets/Scenes/Debug Scene/Scripts/Compounding/compoundingManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b4eb551015dd2ac4b80d367a56c78bc6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
